Simple Pay completed
All the data cleaning and aggregating functions are implemented to the GUI.
Upon processing the Simple Pay files, two files are created for each file type. One is ready for import into the accounting software, the other contains additional information in case invoice number is missing from the mail file.